Assign taxonomy functions

Assign taxonomy functions

Usage

assign_tax(
  analysis_setup,
  asv_abund_matrix,
  retrieve_files = FALSE,
  overwrite_existing = FALSE,
  db_rps10 = "oomycetedb.fasta",
  db_its = "fungidb.fasta",
  db_16S = "bacteriadb.fasta",
  db_other1 = "otherdb1.fasta",
  db_other2 = "otherdb2.fasta"
)

Arguments

analysis_setup

An object containing directory paths and data tables, produced by the prepare_reads function

asv_abund_matrix

The final abundance matrix containing amplified sequence variants

retrieve_files

Logical, TRUE/FALSE whether to copy files from the temp directory to the output directory. Default is FALSE.

overwrite_existing

Logical, indicating whether to remove or overwrite existing files and directories from previous runs. Default is FALSE.

db_rps10

The reference database for the rps10 metabarcode

db_its

The reference database for the ITS metabarcode

db_16S

The SILVA 16S-rRNA reference database provided by the user

db_other1

The reference database for other metabarcode 1 (assumes format is like SILVA DB entries)

db_other2

The reference database for other metabarcode 2 (assumes format is like SILVA DB entries)

Value

Taxonomic assignments of each unique ASV sequence

Details

At this point, 'dada2' function assignTaxonomy is used to assign taxonomy to the inferred ASVs.

Examples

if (FALSE) { # \dontrun{
# Assign taxonomies to ASVs on by metabarcode
analysis_setup <- prepare_reads(
  data_directory = system.file("extdata", package = "demulticoder"),
  output_directory = tempdir(),
  tempdir_path = tempdir(),
  tempdir_id = "demulticoder_run_temp",
  overwrite_existing = TRUE
)
cut_trim(
analysis_setup,
cutadapt_path="/usr/bin/cutadapt",
overwrite_existing = TRUE
)
make_asv_abund_matrix(
analysis_setup,
overwrite_existing = TRUE
)
assign_tax(
analysis_setup,
asv_abund_matrix,
retrieve_files=FALSE,
overwrite_existing = TRUE
)
} # }