Filter ASV abundance matrix and convert to taxmap and phyloseq objects

Filter ASV abundance matrix and convert to taxmap and phyloseq objects

Usage

convert_asv_matrix_to_objs(
  analysis_setup,
  min_read_depth = 0,
  minimum_bootstrap = 0,
  save_outputs = FALSE
)

Arguments

analysis_setup

An object containing directory paths and data tables, produced by the prepare_reads function

min_read_depth

ASV filter parameter. If mean read depth of across all samples is less than this threshold, ASV will be filtered.

minimum_bootstrap

Set threshold for bootstrap support value for taxonomic assignments. Below designated minimum bootstrap threshold, taxnomoic assignments will be set to N/A

save_outputs

Logical, indicating whether to save the resulting phyloseq and taxmap objects. If TRUE, the objects will be saved; if FALSE, they will only be available in the global environment. Default is FALSE.

Value

ASV matrix converted to taxmap object

Examples