Get primer counts for reach sample after primer removal and trimming steps

Get primer counts for reach sample after primer removal and trimming steps

Usage

get_post_trim_hits(primer_data, cutadapt_data, output_directory_path)

Arguments

primer_data

The primer data frame created in orient_primers function

cutadapt_data

directory_data folder with trimmed and filtered reads for each sample

output_directory_path

The path to the directory where resulting files are output

Value

Table of read counts across each sample