Core dada2 function to learn errors and infer ASVs

Core dada2 function to learn errors and infer ASVs

Usage

infer_asvs(
  data_tables,
  my_direction,
  my_primer_pair_id,
  barcode_params,
  output_directory_path
)

Arguments

data_tables

The data tables containing the paths to read files, metadata, and metabarcode information with associated primer sequences

my_direction

Whether primer is in forward or reverse direction

my_primer_pair_id

The specific metabarcode ID

output_directory_path

The path to the directory where resulting files are output

Value

asv_data