Wrapper function for plotQualityProfile function

Wrapper function for plotQualityProfile function

Usage

plot_qc(cutadapt_data, output_directory_path, n = 5e+05)

Arguments

cutadapt_data

directory_data folder with trimmed and filtered reads for each sample

output_directory_path

The path to the directory where resulting files are output

n

(Optional). Default 500,000. The number of records to sample from the fastq file.

Value

Dada2 wrapper function for making quality profiles for each sample