Prepare reads for primer trimming using Cutadapt

This function organizes user-defined directories, metadata and parameter inputs in preparation for Cutadapt to remove primers from short read metabarcoding sequence data. Afterwards, subsequent functions can be used to simplify core DADA2 ASV and taxonomic assignment steps.

prepare_reads(
  data_directory = "data",
  output_directory = "output",
  tempdir_path = NULL,
  tempdir_id = "demulticoder_run",
  maxN = 0,
  multithread = FALSE,
  overwrite_existing = FALSE
)

Arguments

data_directory

User-specified directory path where user has placed raw FASTQ (forward and reverse reads), metadata.csv, and primerinfo_params.csv files. Default is "data".

output_directory

User-specified directory for outputs. Default is "output".

tempdir_path

Path to a temporary directory. If NULL, a temporary directory path will be identified using tempdir() command.

tempdir_id

ID for temporary directories. Default is "demulticoder_run". User can provide any helpful ID, whether it be a date or specific name for the run.

maxN

(Optional). Default 0. After truncation, sequences with more than maxN Ns will be discarded. Note that dada does not allow Ns.

multithread

(Optional). Default is FALSE. If TRUE, input files are filtered in parallel via mclapply. If an integer is provided, it is passed to the mc.cores argument of mclapply. Note that the parallelization here is by forking, and each process is loading another fastq file into memory. This option is ignored in Windows, as Windows does not support forking, with mc.cores set to 1. If memory is an issue, execute in a clean environment and reduce the chunk size n and/or the number of threads.

overwrite_existing

Logical, indicating whether to remove or overwrite existing files and directories from previous runs.

Value

A list containing data tables, including metadata, primer sequences, paths for trimming reads, and user-defined parameters.

Examples