Wrapper function for core DADA2 filter and trim function for first filtering step

Wrapper function for core DADA2 filter and trim function for first filtering step

Usage

remove_ns(fastq_data, multithread, temp_directory_path)

Arguments

fastq_data

A data frame with the fastq file paths, the direction of the sequences, and names of sequences metadata, and primer_info files

multithread

(Optional). Default is FALSE. If TRUE, input files are filtered in parallel via mclapply. If an integer is provided, it is passed to the mc.cores argument of mclapply. Note that the parallelization here is by forking, and each process is loading another fastq file into memory. This option is ignored in Windows, as Windows does not support forking, with mc.cores set to 1. If memory is an issue, execute in a clean environment and reduce the chunk size n and/or the number of threads.

temp_directory_path

User-defined temporary directory to output unfiltered, trimmed, and filtered read directories throughout the workflow

metadata

A metadata containing the concatenated metadata and primer data

Value

Return prefiltered reads with no Ns