Set up directory paths for subsequent analyses

This function sets up the directory paths for subsequent analyses. It checks whether the specified output directories exist or creates them if they don't. The function also provides paths to primer and metadata files within the data directory.

Usage

setup_directories(
  data_directory = "data",
  output_directory = "output",
  tempdir_path = NULL,
  tempdir_id = "demulticoder_run"
)

Arguments

data_directory

User-specified directory path where the user has placed raw FASTQ (forward and reverse reads), metadata.csv, and primerinfo_params.csv files. Default is "data".

output_directory

User-specified directory for outputs. Default is "output".

tempdir_path

Path to a temporary directory. If NULL, a temporary directory path will be identified using the tempdir() command.

tempdir_id

ID for temporary directories. Default is "demulticoder_run". The user can provide any helpful ID, whether it be a date or specific name for the run.

Value

A list with paths for data, output, temporary directories, primer, and metadata files.