Assign taxonomy functions

Assign taxonomy functions

assign_tax(
  analysis_setup,
  asv_abund_matrix,
  tryRC = FALSE,
  verbose = FALSE,
  multithread = FALSE,
  retrieve_files = FALSE,
  db_rps10 = "oomycetedb.fasta",
  db_its = "fungidb.fasta",
  db_16s = "bacteriadb.fasta",
  db_other1 = "otherdb1.fasta",
  db_other2 = "otherdb2.fasta",
  overwrite_existing = FALSE
)

Arguments

analysis_setup

A list containing directory paths and data tables, produced by the prepare_reads function.

asv_abund_matrix

ASV abundance matrix.

tryRC

Whether to try reverse complementing sequences during taxonomic assignment.

verbose

Logical, indicating whether to display verbose output.

multithread

Logical, indicating whether to use multithreading.

retrieve_files

Specify TRUE/FALSE whether to copy files from the temp directory to the output directory.

overwrite_existing

Logical, indicating whether to remove or overwrite existing files and directories from previous runs.

Value

Taxonomic assignments of each unique ASV sequence

Examples